# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Courier Services OpenApi Documentation",
        "description": "L5 Swagger OpenApi description",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/courier-services/api/v1",
            "description": "CMS Courier Services V1"
        }
    ],
    "paths": {
...
        "/admin/area/destroy/{id}": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "delete area",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "403": {
                        "description": "Unauthorized Access"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "apiAuth": [],
                        "PLATFORM": [],
                        "CPNYAPIKEY": []
                    }
                ]
            }
        },
        "/admin/area/activate/{id}": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "change area status",
                "operationId": "activate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "403": {
                        "description": "Unauthorized Access"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "apiAuth": [],
                        "PLATFORM": [],
                        "CPNYAPIKEY": []
                    }
                ]
            }
        },
        "/admin/area/deactivate/{id}": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "change area status",
                "operationId": "deactivate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "403": {
                        "description": "Unauthorized Access"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "apiAuth": [],
                        "PLATFORM": [],
                        "CPNYAPIKEY": []
                    }
                ]
            }
        },
...
        "/admin/inquiry_area/destroy/{id}": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "delete",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "403": {
                        "description": "Unauthorized Access"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "apiAuth": [],
                        "PLATFORM": [],
                        "CPNYAPIKEY": []
                    }
                ]
            }
        },
        "/admin/inquiry_area/activate/{id}": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "change status",
                "operationId": "activate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "403": {
                        "description": "Unauthorized Access"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "apiAuth": [],
                        "PLATFORM": [],
                        "CPNYAPIKEY": []
                    }
                ]
            }
        },
        "/admin/inquiry_area/deactivate/{id}": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "change status",
                "operationId": "deactivate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "403": {
                        "description": "Unauthorized Access"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "apiAuth": [],
                        "PLATFORM": [],
                        "CPNYAPIKEY": []
                    }
                ]
            }
        },
...
        "/admin/premises/destroy/{id}": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "delete premises",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "403": {
                        "description": "Unauthorized Access"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "apiAuth": [],
                        "PLATFORM": [],
                        "CPNYAPIKEY": []
                    }
                ]
            }
        },
        "/admin/premises/activate/{id}": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "change premises status",
                "operationId": "activate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "403": {
                        "description": "Unauthorized Access"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "apiAuth": [],
                        "PLATFORM": [],
                        "CPNYAPIKEY": []
                    }
                ]
            }
        },
        "/admin/premises/deactivate/{id}": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "change premises status",
                "operationId": "deactivate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "403": {
                        "description": "Unauthorized Access"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "apiAuth": [],
                        "PLATFORM": [],
                        "CPNYAPIKEY": []
                    }
                ]
            }
        },
...
        "/admin/premises_type/lov": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "get premises type record",
                "operationId": "getStatesRecord",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "403": {
                        "description": "Unauthorized Access"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "apiAuth": [],
                        "PLATFORM": [],
                        "CPNYAPIKEY": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "PLATFORM": {
                "type": "apiKey",
                "name": "PLATFORM",
                "in": "header"
            },
            "CPNYAPIKEY": {
                "type": "apiKey",
                "name": "CPNYAPIKEY",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}